Fix default start date in mission form
